Add Team permissions and update unit tests accordingly

+++ src/app/Services/TeamService.php
@@ -5,6 +5,7 @@ namespace App\Services;
 use App\Http\Requests\Team\TeamStoreRequest;
 use App\Http\Requests\Team\TeamUpdateRequest;
 use App\Models\Team;
+use Illuminate\Support\Facades\Gate;
 
 class TeamService
 {
@@ -18,6 +19,8 @@ class TeamService
 
     public function store(TeamStoreRequest $request): Team
     {
+        Gate::authorize('create', Team::class);
+
         $validated = $request->validated();
 
         $team = Team::create([
@@ -31,7 +34,10 @@ class TeamService
 
     public function show(string $id): Team
     {
-        return Team::findOrFail($id);
+        $team = Team::findOrFail($id);
+        Gate::authorize('view', $team);
+
+        return $team;
     }
 
     public function update(TeamUpdateRequest $request, string $id): Team
@@ -39,6 +45,7 @@ class TeamService
         $validated = $request->validated();
         $team = $this->show($id);
 
+        Gate::authorize('update', $team);
         $team->update($validated);
 
         return $team;
@@ -47,6 +54,8 @@ class TeamService
     public function delete(string $id): void
     {
         $team = Team::findOrFail($id);
+
+        Gate::authorize('delete', $team);
 
         $team->delete();
     }
